requirements.txt
alembic==1.13.1
Flask==3.0.3
Flask_JWT_Extended==4.6.0
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
pyodbc==5.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
